--- .github/workflows/pytest.yml
@@ -1,91 +1,0 @@
-name: Tests
-
-on: [push, pull_request]
-
-jobs:
-  ubuntu:
-
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: [3.6, 3.7, 3.8]
-        use-conda: [true, false]
-        use-dist: [false]
-        include:
-          - python-version: 3.8
-            code-cov: true
-          - python-version: 3.7
-            use-conda: false
-            use-dist: true
-      fail-fast:  false
-
-    steps:
-    - uses: actions/checkout@v2
-    - name: Setup Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v2
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Conda Install test dependencies 
-      if: matrix.use-conda == true
-      run: |
-        # Miniconda is available in $CONDA env var
-        $CONDA/bin/conda create -n testenv --yes pip wheel gxx_linux-64 gcc_linux-64 swig python=${{ matrix.python-version }}
-        $CONDA/envs/testenv/bin/python3 -m pip install --upgrade pip
-        $CONDA/envs/testenv/bin/pip3 install -e .[test]
-    - name: Install test dependencies 
-      if: matrix.use-conda == false && matrix.use-dist == false
-      run: |
-        python -m pip install --upgrade pip
-        pip install -e .[test]
-        sudo apt-get update
-    - name: Dist Install test dependencies 
-      if: matrix.use-conda == false && matrix.use-dist == true
-      run: |
-        python -m pip install --upgrade pip
-        sudo apt-get update
-        # We need to install for the dependencies, like pytest
-        pip install -e .[test]
-        pip uninstall --yes pynisher
-        python setup.py sdist
-        last_dist=$(ls -t dist/pynisher-*.tar.gz | head -n 1)
-        pip install $last_dist
-    - name: Store repository status
-      id: status-before
-      run: |
-        echo "::set-output name=BEFORE::$(git status --porcelain -b)"
-    - name: Conda Run tests
-      if: matrix.use-conda == true
-      run: |
-        # We activate conda as metalearning uses python directly, so we need
-        # to change the default python
-        export PATH="$CONDA/envs/testenv/bin:$PATH"
-        if [ ${{ matrix.code-cov }} ]; then codecov='--cov=pynisher --cov-report=xml'; fi
-        # Forked is imperative as spawn creates <defunct> children
-        # This processes cannot be killed unless killing the parent
-        # The forked isolates each check to it's own process
-        $CONDA/envs/testenv/bin/python3 -m pytest --forked --durations=20 -v $codecov test
-    - name: Run tests
-      if: matrix.use-conda == false
-      run: |
-        if [ ${{ matrix.code-cov }} ]; then codecov='--cov=pynisher --cov-report=xml'; fi
-        # Forked is imperative as spawn creates <defunct> children
-        # This processes cannot be killed unless killing the parent
-        # The forked isolates each check to it's own process
-        pytest --forked --durations=20 -v $codecov test
-    - name: Check for files left behind by test
-      if: ${{ always() }}
-      run: |
-        before="${{ steps.status-before.outputs.BEFORE }}"
-        after="$(git status --porcelain -b)"
-        if [[ "$before" != "$after" ]]; then
-            echo "git status from before: $before"
-            echo "git status from after: $after"
-            echo "Not all generated files have been deleted!"
-            exit 1
-        fi
-    - name: Upload coverage
-      if: matrix.code-cov && always()
-      uses: codecov/codecov-action@v1
-      with:
-        fail_ci_if_error: true
-        verbose: true
