+++ test/test_pynisher.py
@@ -1,25 +1,37 @@
 #! /bin/python
-import logging
-import multiprocessing
-import os
-import pytest
-import signal
 import time
+import multiprocessing
 import unittest
 import unittest.mock
+import os
+import signal
+import logging
 
 import psutil
 
 import pynisher
 
+try:
+    import sklearn # noqa
+
+    is_sklearn_available = True
+except ImportError:
+    print("Scikit Learn was not found!")
+    is_sklearn_available = False
+
+context = os.environ.get('CONTEXT', 'fork')
+expected_children = {
+    'fork': 0,
+    'spawn': 1,
+    'forkserver': 2,
+}
+print(f"Using context = {context}")
+all_tests = 1
 logger = multiprocessing.log_to_stderr()
 logger.setLevel(logging.WARNING)
 
 
 # TODO: add tests with large return value to test for deadlock!
-
-# Functions below have to be at the top of the file to be pick-able
-# for multiprocessing
 
 def rogue_subprocess():
     pid = os.getpid()
@@ -87,11 +99,6 @@ def crash_with_exception(exception):
     raise exception
 
 
-def crash_while_read_file():
-    with open('some_non_existing_files_to_cause_crash'):
-        pass
-
-
 def return_big_array(num_elements):
     return ([1] * num_elements)
 
@@ -100,6 +107,17 @@ def cpu_usage():
     i = 1
     while True:
         i += 1
+
+
+def nested_pynisher(level=2, cputime=5, walltime=5, memlimit=10e24, increment=-1, grace_period=1):
+    print("this is level {}".format(level))
+    if level == 0:
+        spawn_rogue_subprocess(10)
+    else:
+        func = pynisher.enforce_limits(mem_in_mb=memlimit, cpu_time_in_s=cputime, wall_time_in_s=walltime,
+                                       context=multiprocessing.get_context(context),
+                                       grace_period_in_s=grace_period)(nested_pynisher)
+        func(level - 1, None, walltime + increment, memlimit, increment)
 
 
 def print_and_sleep(t):
@@ -113,31 +131,15 @@ def print_and_fail():
     raise RuntimeError()
 
 
-def nested_pynisher(level=2, cputime=5, walltime=5, memlimit=10e24,
-                    increment=-1, grace_period=1, context=None):
-    print("this is level {}".format(level))
-    if level == 0:
-        spawn_rogue_subprocess(10)
-    else:
-        func = pynisher.enforce_limits(
-            mem_in_mb=memlimit, cpu_time_in_s=cputime,
-            wall_time_in_s=walltime,
-            grace_period_in_s=grace_period,
-            context=context,
-        )(nested_pynisher)
-        func(level=level - 1,
-             cputime=None,
-             walltime=walltime + increment,
-             memlimit=memlimit,
-             increment=increment,
-             context=context,
-             )
+def crash_while_read_file():
+    with open('some_non_existing_files_to_cause_crash'):
+        pass
 
 
-@pytest.mark.parametrize("context", ['fork', 'spawn', 'forkserver'])
-class TestLimitResources:
-    @pytest.mark.parametrize("memory", [1, 2, 4, 8, 16])
-    def test_success(self, context, memory):
+class test_limit_resources_module(unittest.TestCase):
+
+    @unittest.skipIf(not all_tests, "skipping successful tests")
+    def test_success(self):
 
         print("Testing unbounded function call which have to run through!")
         local_mem_in_mb = None
@@ -145,20 +147,18 @@ class TestLimitResources:
         local_cpu_time_in_s = None
         local_grace_period = None
 
-        wrapped_function = pynisher.enforce_limits(
-            mem_in_mb=local_mem_in_mb,
-            wall_time_in_s=local_wall_time_in_s,
-            cpu_time_in_s=local_cpu_time_in_s,
-            grace_period_in_s=local_grace_period,
-            context=multiprocessing.get_context(context),
-        )(simulate_work)
+        wrapped_function = pynisher.enforce_limits(mem_in_mb=local_mem_in_mb, wall_time_in_s=local_wall_time_in_s,
+                                                   context=multiprocessing.get_context(context),
+                                                   cpu_time_in_s=local_cpu_time_in_s,
+                                                   grace_period_in_s=local_grace_period)(simulate_work)
 
-        assert (memory, 0, 0) == wrapped_function(memory, 0, 0)
-        assert wrapped_function.exit_status == 0
-        assert wrapped_function.exitcode == 0
+        for mem in [1, 2, 4, 8, 16]:
+            self.assertEqual((mem, 0, 0), wrapped_function(mem, 0, 0))
+            self.assertEqual(wrapped_function.exit_status, 0)
+            self.assertEqual(wrapped_function.exitcode, 0)
 
-    @pytest.mark.parametrize("memory", [1024, 2048, 4096])
-    def test_out_of_memory(self, context, memory):
+    @unittest.skipIf(not all_tests, "skipping out_of_memory test")
+    def test_out_of_memory(self):
         print("Testing memory constraint.")
         local_mem_in_mb = 32
         local_wall_time_in_s = None
@@ -166,145 +166,110 @@ class TestLimitResources:
         local_grace_period = None
 
         wrapped_function = pynisher.enforce_limits(
-            mem_in_mb=local_mem_in_mb,
-            wall_time_in_s=local_wall_time_in_s,
-            cpu_time_in_s=local_cpu_time_in_s,
-            grace_period_in_s=local_grace_period,
+            mem_in_mb=local_mem_in_mb, wall_time_in_s=local_wall_time_in_s,
             context=multiprocessing.get_context(context),
-        )(simulate_work)
+            cpu_time_in_s=local_cpu_time_in_s,
+            grace_period_in_s=local_grace_period)(simulate_work)
 
-        assert wrapped_function(memory, 0, 0) is None, "{}/{}".format(wrapped_function.result,
-                                                                      wrapped_function.exit_status)
-        assert wrapped_function.exit_status == pynisher.MemorylimitException
-        assert wrapped_function.exitcode == 0
+        for mem in [1024, 2048, 4096]:
+            self.assertIsNone(wrapped_function(mem, 0, 0))
+            self.assertEqual(wrapped_function.exit_status, pynisher.MemorylimitException)
+            self.assertEqual(wrapped_function.exitcode, 0)
 
-    @pytest.mark.parametrize("memory", [1, 10])
-    def test_time_out(self, context, memory):
+    @unittest.skipIf(not all_tests, "skipping time_out test")
+    def test_time_out(self):
         print("Testing wall clock time constraint.")
         local_mem_in_mb = None
         local_wall_time_in_s = 1
         local_cpu_time_in_s = None
         local_grace_period = None
 
-        wrapped_function = pynisher.enforce_limits(
-            mem_in_mb=local_mem_in_mb,
-            wall_time_in_s=local_wall_time_in_s,
-            cpu_time_in_s=local_cpu_time_in_s,
-            grace_period_in_s=local_grace_period,
-            context=multiprocessing.get_context(context),
-        )(simulate_work)
+        wrapped_function = pynisher.enforce_limits(mem_in_mb=local_mem_in_mb, wall_time_in_s=local_wall_time_in_s,
+                                                   context=multiprocessing.get_context(context),
+                                                   cpu_time_in_s=local_cpu_time_in_s,
+                                                   grace_period_in_s=local_grace_period)(simulate_work)
 
-        assert wrapped_function(memory, 10, 0) is None
-        assert wrapped_function.exit_status == pynisher.TimeoutException, str(
-            wrapped_function.result)
-        # Apparently, the exit code here is not deterministic (so far only PYthon 3.6)
-        assert wrapped_function.exitcode in (-15, 0)
+        for mem in range(1, 10):
+            self.assertIsNone(wrapped_function(mem, 10, 0))
+            self.assertEqual(wrapped_function.exit_status, pynisher.TimeoutException, str(wrapped_function.result))
+            # Apparently, the exit code here is not deterministic (so far only PYthon 3.6)
+            self.assertIn(wrapped_function.exitcode, (-15, 0))
 
-    @pytest.mark.parametrize("processes", [2, 15, 50, 100, 250])
-    def test_num_processes(self, context, processes):
+    @unittest.skipIf(not all_tests, "skipping too many processes test")
+    def test_num_processes(self):
         print("Testing number of processes constraint.")
         local_mem_in_mb = None
         local_num_processes = 1
         local_wall_time_in_s = None
         local_grace_period = None
 
-        wrapped_function = pynisher.enforce_limits(
-            mem_in_mb=local_mem_in_mb, wall_time_in_s=local_wall_time_in_s,
-            num_processes=local_num_processes,
-            grace_period_in_s=local_grace_period,
-            context=multiprocessing.get_context(context),
-        )(simulate_work)
+        wrapped_function = pynisher.enforce_limits(mem_in_mb=local_mem_in_mb, wall_time_in_s=local_wall_time_in_s,
+                                                   context=multiprocessing.get_context(context),
+                                                   num_processes=local_num_processes,
+                                                   grace_period_in_s=local_grace_period)(simulate_work)
 
-        assert wrapped_function(0, 0, processes) is None
-        assert wrapped_function.exit_status == pynisher.SubprocessException
-        assert wrapped_function.exitcode == 0
+        for processes in [2, 15, 50, 100, 250]:
+            self.assertIsNone(wrapped_function(0, 0, processes))
+            self.assertEqual(wrapped_function.exit_status, pynisher.SubprocessException)
+            self.assertEqual(wrapped_function.exitcode, 0)
 
-    def test_crash_unexpectedly(self, context):
+    @unittest.skipIf(not all_tests, "skipping unexpected signal test")
+    def test_crash_unexpectedly(self):
         print("Testing an unexpected signal simulating a crash.")
         wrapped_function = pynisher.enforce_limits(
             context=multiprocessing.get_context(context),
         )(crash_unexpectedly)
-        assert wrapped_function(signal.SIGQUIT) is None
-        assert wrapped_function.exit_status == pynisher.SignalException
-        assert wrapped_function.exitcode == 0
+        self.assertIsNone(wrapped_function(signal.SIGQUIT))
+        self.assertEqual(wrapped_function.exit_status, pynisher.SignalException)
+        self.assertEqual(wrapped_function.exitcode, 0)
 
-    def test_high_cpu_percentage(self, context):
+    @unittest.skipIf(not all_tests, "skipping unexpected signal test")
+    def test_high_cpu_percentage(self):
         print("Testing cpu time constraint.")
-        # From the multiprocessing documentation:
-        # Starting a process using this method ('spawn') is rather slow
-        # compared to using fork or forkserver.
-        # Increasing cpu_time_in_s from 2 to 4 seconds
-        cpu_time_in_s = 4
+        cpu_time_in_s = 2
         grace_period = 1
-        wrapped_function = pynisher.enforce_limits(
-            cpu_time_in_s=cpu_time_in_s, grace_period_in_s=grace_period,
-            context=multiprocessing.get_context(context),
-        )(cpu_usage)
+        wrapped_function = pynisher.enforce_limits(cpu_time_in_s=cpu_time_in_s,
+                                                   context=multiprocessing.get_context(context),
+                                                   grace_period_in_s=grace_period)(
+            cpu_usage)
 
-        assert wrapped_function() is None
-        assert wrapped_function.exit_status == pynisher.CpuTimeoutException
-        assert wrapped_function.exitcode == 0
+        self.assertIsNone(wrapped_function())
+        self.assertEqual(wrapped_function.exit_status, pynisher.CpuTimeoutException)
+        self.assertEqual(wrapped_function.exitcode, 0)
 
-    @pytest.mark.parametrize("num_elements", [4, 16, 64, 256, 1024, 4096, 16384, 65536, 262144])
-    def test_big_return_data(self, context, num_elements):
+    @unittest.skipIf(not all_tests, "skipping big data test")
+    def test_big_return_data(self):
         print("Testing big return values")
         wrapped_function = pynisher.enforce_limits(
             context=multiprocessing.get_context(context),
         )(return_big_array)
 
-        bla = wrapped_function(num_elements)
-        assert len(bla) == num_elements
-        assert wrapped_function.exitcode == 0
+        for num_elements in [4, 16, 64, 256, 1024, 4096, 16384, 65536, 262144]:
+            bla = wrapped_function(num_elements)
+            self.assertEqual(len(bla), num_elements)
+            self.assertEqual(wrapped_function.exitcode, 0)
 
-    def test_kill_subprocesses(self, context):
-
-        # Work around to address the fact that the main process had left children
-        # when running all test in pytest
-        initial_recursive_children = len(psutil.Process().children(recursive=True))
+    @unittest.skipIf(not all_tests, "skipping subprocess changing process group")
+    def test_kill_subprocesses(self):
         wrapped_function = pynisher.enforce_limits(
-            wall_time_in_s=1,
             context=multiprocessing.get_context(context),
-        )(spawn_rogue_subprocess)
+            wall_time_in_s=1)(spawn_rogue_subprocess)
         wrapped_function(5)
 
         time.sleep(1)
         p = psutil.Process()
-        if context == 'spawn':
-            # The parent process starts a fresh python interpreter process.
-            assert len(p.children(recursive=True)) - initial_recursive_children == 1, "current={} child={} parent={} initial={}".format(
-                p,
-                p.children(recursive=True),
-                os.getppid(),
-                initial_recursive_children
-            )
-        elif context == 'forkserver':
-            # When the program starts and selects the forkserver start method,
-            # a server process is started. From then on, whenever a new process
-            # is needed, the parent process connects to the server and requests
-            # that it fork a new process.
-            assert len(p.children(recursive=True)) - initial_recursive_children == 2, "current={} child={} parent={} initial={}".format(
-                p,
-                p.children(recursive=True),
-                os.getppid(),
-                initial_recursive_children
-            )
-        else:
-            assert len(p.children(recursive=True)) - initial_recursive_children == 0, "current={} child={} parent={} initial={}".format(
-                p,
-                p.children(recursive=True),
-                os.getppid(),
-                initial_recursive_children
-            )
-        assert wrapped_function.exitcode == -15
+        self.assertEqual(len(p.children(recursive=True)), expected_children[context])
+        self.assertEqual(wrapped_function.exitcode, -15)
 
-    def test_busy_in_C_library(self, context):
+    @unittest.skipIf(not is_sklearn_available, "test requires scikit learn")
+    @unittest.skipIf(not all_tests, "skipping fitting an SVM to see how C libraries are handles")
+    def test_busy_in_C_library(self):
 
         global logger
 
         wrapped_function = pynisher.enforce_limits(
-            wall_time_in_s=2,
             context=multiprocessing.get_context(context),
-        )(svm_example)
+            wall_time_in_s=2)(svm_example)
 
         start = time.time()
         wrapped_function(16384, 128)
@@ -312,128 +277,121 @@ class TestLimitResources:
 
         time.sleep(1)
         p = psutil.Process()
-        if context == 'spawn':
-            # The parent process starts a fresh python interpreter process.
-            assert len(p.children(recursive=True)) == 1
-        elif context == 'forkserver':
-            # When the program starts and selects the forkserver start method,
-            # a server process is started. From then on, whenever a new process
-            # is needed, the parent process connects to the server and requests
-            # that it fork a new process.
-            assert len(p.children(recursive=True)) == 2
-        else:
-            assert len(p.children(recursive=True)) == 0
-        # In github actions, I have seen up to 2.2 due to delay
-        # relaxing time a bit from 2.1->2.3
-        assert duration <= 2.3
-        assert wrapped_function.exitcode == -15
+        self.assertEqual(len(p.children(recursive=True)), expected_children[context])
+        self.assertTrue(duration <= 2.1)
+        self.assertEqual(wrapped_function.exitcode, -15)
+        self.assertLess(duration, 2.1)
 
-    def test_liblinear_svc(self, context, logger_mock):
-        # Fitting an SVM to see how C libraries are handles
+    @unittest.skipIf(not is_sklearn_available, "test requires scikit learn")
+    @unittest.skipIf(not all_tests, "skipping fitting an SVM to see how C libraries are handles")
+    @unittest.skipIf(context != 'fork', "I don't think this function works even in fork. It produces a EOFError. But in spawn/forkserver it last less than expected time.")
+    def test_liblinear_svc(self):
 
         global logger
 
         time_limit = 2
         grace_period = 1
+        logger_mock = unittest.mock.Mock()
 
-        wrapped_function = pynisher.enforce_limits(
-            cpu_time_in_s=time_limit, mem_in_mb=None,
-            grace_period_in_s=grace_period, logger=logger,
-            context=multiprocessing.get_context(context),
-        )
+        wrapped_function = pynisher.enforce_limits(cpu_time_in_s=time_limit, mem_in_mb=None,
+                                                   context=multiprocessing.get_context(context),
+                                                   grace_period_in_s=grace_period, logger=logger)
         wrapped_function.logger = logger_mock
         wrapped_function = wrapped_function(svc_example)
         start = time.time()
         wrapped_function(16384, 10000)
         duration = time.time() - start
 
-        # Need more time for spawn processes -- the doc
-        # say they are slow!
-        time.sleep(5)
+        time.sleep(1)
         p = psutil.Process()
+        self.assertEqual(len(p.children(recursive=True)), expected_children[context])
+        self.assertEqual(logger_mock.debug.call_count, 2)
+        self.assertEqual(logger_mock.debug.call_args_list[0][0][0],
+                         'Function called with argument: (16384, 10000), {}')
+        self.assertEqual(logger_mock.debug.call_args_list[1][0][0],
+                         'Your function call closed the pipe prematurely -> '
+                         'Subprocess probably got an uncatchable signal.')
+        # self.assertEqual(wrapped_function.exit_status, pynisher.CpuTimeoutException)
+        self.assertGreater(duration, time_limit - 0.1)
+        self.assertLess(duration, time_limit + grace_period + 0.1)
+        self.assertEqual(wrapped_function.exitcode, -9)
 
-        assert logger_mock.debug.call_count == 2
-        assert logger_mock.debug.call_args_list[0][0][0] == 'Function called with argument: (16384, 10000), {}'
-        assert logger_mock.debug.call_args_list[1][0][0] == 'Your function call closed the pipe prematurely -> Subprocess probably got an uncatchable signal.'
-        # In the case of spawn/forkserver
-        if context == 'spawn':
-            # The parent process starts a fresh python interpreter process.
-            assert len(p.children(recursive=True)) == 1
-        elif context == 'forkserver':
-            # When the program starts and selects the forkserver start method,
-            # a server process is started. From then on, whenever a new process
-            # is needed, the parent process connects to the server and requests
-            # that it fork a new process.
-            assert len(p.children(recursive=True)) == 2
-        else:
-            assert len(p.children(recursive=True)) == 0
-            # self.assertEqual(wrapped_function.exit_status, pynisher.CpuTimeoutException)
-            assert duration > (time_limit - 0.1)
-            assert duration < (time_limit + grace_period + 0.1)
-        assert wrapped_function.exitcode == -9
-
-    def test_nesting(self, context):
-
-        if context in ['spawn', 'forkserver']:
-            pytest.skip("Cannot perform nested functions under spawn/forserved due to EOFError")
+    @unittest.skipIf(not all_tests, "skipping nested pynisher test")
+    def test_nesting(self):
 
         tl = 2  # time limit
         gp = 1  # grace period
 
         start = time.time()
-        nested_pynisher(
-            level=2, cputime=2, walltime=2, memlimit=None, increment=1, grace_period=gp,
-            context=multiprocessing.get_context(context),
-        )
+        nested_pynisher(level=2, cputime=2, walltime=2, memlimit=None, increment=1, grace_period=gp)
         duration = time.time() - start
         print(duration)
 
         time.sleep(1)
         p = psutil.Process()
-        assert len(p.children(recursive=True)) == 0
-        assert duration > tl - 0.1
-        assert duration < tl + gp + 0.1
+        self.assertEqual(len(p.children(recursive=True)), expected_children[context])
+        self.assertGreater(duration, tl - 0.1)
+        self.assertLess(duration, tl + gp + 0.1)
 
-    def test_capture_output(self, context):
+    @unittest.skipIf(not all_tests, "skipping capture stdout test")
+    def test_capture_output(self):
         print("Testing capturing of output.")
         global logger
 
         time_limit = 2
         grace_period = 1
 
-        wrapped_function = pynisher.enforce_limits(
-            wall_time_in_s=time_limit, mem_in_mb=None,
-            grace_period_in_s=grace_period, logger=logger, capture_output=True,
-            context=multiprocessing.get_context(context),
-        )(print_and_sleep)
+        wrapped_function = pynisher.enforce_limits(wall_time_in_s=time_limit, mem_in_mb=None,
+                                                   context=multiprocessing.get_context(context),
+                                                   grace_period_in_s=grace_period, logger=logger, capture_output=True)(
+            print_and_sleep)
 
         wrapped_function(5)
 
-        assert '0' in wrapped_function.stdout
-        assert wrapped_function.stderr == ''
-        assert wrapped_function.exitcode == 0
+        self.assertTrue('0' in wrapped_function.stdout)
+        self.assertEqual(wrapped_function.stderr, '')
+        self.assertEqual(wrapped_function.exitcode, 0)
 
-        wrapped_function = pynisher.enforce_limits(
-            wall_time_in_s=time_limit, mem_in_mb=None,
-            grace_period_in_s=grace_period, logger=logger, capture_output=True,
-            context=multiprocessing.get_context(context),
-        )(print_and_fail)
+        wrapped_function = pynisher.enforce_limits(wall_time_in_s=time_limit, mem_in_mb=None,
+                                                   context=multiprocessing.get_context(context),
+                                                   grace_period_in_s=grace_period, logger=logger, capture_output=True)(
+            print_and_fail)
 
         wrapped_function()
 
-        assert '0' in wrapped_function.stdout
-        assert 'RuntimeError' in wrapped_function.stderr
-        assert wrapped_function.exitcode == 1
+        self.assertIn('0', wrapped_function.stdout)
+        self.assertIn('RuntimeError', wrapped_function.stderr)
+        self.assertEqual(wrapped_function.exitcode, 1)
 
-        assert wrapped_function.exitcode == 0
+    def test_too_little_memory(self):
+        # Test what happens if the target process does not have a sufficiently high memory limit
 
-    def test_raise(self, context):
-        # As above test does not reliably work on travis-ci,
-        # this test checks whether an OSError's error code is properly read out
+        # 2048 MB
+        dummy_content = [42.] * ((1024 * 2048) // 8) # noqa
+
+        wrapped_function = pynisher.enforce_limits(mem_in_mb=1,
+                                                   context=multiprocessing.get_context(context),
+                                                   )(simulate_work)
+
+        wrapped_function(size_in_mb=1000, wall_time_in_s=10, num_processes=1,
+                         dummy_content=dummy_content)
+
+        self.assertIsNone(wrapped_function.result)
+        # The following is a bit weird, on my local machine I get a SubprocessException, but on
+        # travis-ci I get a MemoryLimitException
+        self.assertIn(wrapped_function.exit_status,
+                      (pynisher.SubprocessException, pynisher.MemorylimitException))
+        # This is triggered on my local machine, but not on travis-ci
+        if wrapped_function.exit_status == pynisher.SubprocessException:
+            self.assertEqual(wrapped_function.os_errno, 12)
+        self.assertEqual(wrapped_function.exitcode, 0)
+
+    def test_raise(self):
+        # As above test does not reliably work on travis-ci, this test checks whether an
+        # OSError's error code is properly read out
         wrapped_function = pynisher.enforce_limits(
-            mem_in_mb=1000,
             context=multiprocessing.get_context(context),
-        )(crash_while_read_file)
+            mem_in_mb=1000)(crash_while_read_file)
         wrapped_function.logger = unittest.mock.Mock()
 
         # Emulate OS error via file access, as OSError object is reconstructed
@@ -442,39 +400,12 @@ class TestLimitResources:
         # 2 - Misuse of shell builtins (according to Bash documentation)
         wrapped_function()
 
-        assert wrapped_function.result is None
-        assert wrapped_function.exit_status == pynisher.SubprocessException
+        self.assertIsNone(wrapped_function.result)
+        self.assertEqual(wrapped_function.exit_status, pynisher.SubprocessException)
         if wrapped_function.exit_status == pynisher.SubprocessException:
-            assert wrapped_function.os_errno == 2
-        assert wrapped_function.exitcode == 0
+            self.assertEqual(wrapped_function.os_errno, 2)
+        self.assertEqual(wrapped_function.exitcode, 0)
 
 
-# @pytest.mark.parametrize("context", ['fork', 'spawn', 'forkserver'])
-class TestIsolatedLimitResources:
-    """
-    To test out little memory, exclude other fixture/data from the below test
-    """
-    # def test_too_little_memory(self, context):
-    def test_too_little_memory(self):
-        # Test what happens if the target process does not have a
-        # sufficiently high memory limit
-
-        # 2048 MB
-        dummy_content = [42.] * ((1024 * 2048) // 8) # noqa
-
-        wrapped_function = pynisher.enforce_limits(
-            mem_in_mb=1,
-            # context=multiprocessing.get_context(context),
-        )(simulate_work)
-
-        wrapped_function(size_in_mb=1000, wall_time_in_s=10, num_processes=1,
-                         dummy_content=dummy_content)
-
-        assert wrapped_function.result is None
-        # The following is a bit weird, on my local machine I get a SubprocessException,
-        # but on travis-ci I get a MemoryLimitException
-        assert wrapped_function.exit_status in (pynisher.SubprocessException,
-                                                pynisher.MemorylimitException)
-        # This is triggered on my local machine, but not on travis-ci
-        if wrapped_function.exit_status == pynisher.SubprocessException:
-            assert wrapped_function.os_errno == 12
+if __name__ == '__main__':
+    unittest.main()
